+++ app.js
@@ -1,96 +1,25 @@
 const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
-const { listContacts, getContactById, addContact, removeContact, updateContact, validateContact } = require('./routes/api/contacts');
+
+const contactsRouter = require('./routes/api/contacts')
 
 const app = express()
 
-const PORT = process.env.PORT || 3000;
+const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'
 
-app.use(express.json())
-app.use(logger('combined'))
+app.use(logger(formatsLogger))
 app.use(cors())
+app.use(express.json())
 
+app.use('/api/contacts', contactsRouter)
 
-app.get('/api/contacts', async (req, res) => {
-  try {
-    const contacts = await listContacts();
-    res.status(200).json(contacts);
-  } catch (error) {
-    console.error(error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-});
+app.use((req, res) => {
+  res.status(404).json({ message: 'Not found' })
+})
 
-app.get('/api/contacts/:id', async (req, res) => {
-  const { id } = req.params;
-  try {
-    const contact = await getContactById(id);
-    if (contact) {
-      res.status(200).json(contact);
-    } else {
-      res.status(404).json({ message: 'Not found' });
-    }
-  } catch (error) {
-    console.error(error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-});
-
-app.post('/api/contacts', async (req, res) => {
-  const { name, email, phone } = req.body;
-
-  const validationResult = validateContact({ name, email, phone });
-
-  if (validationResult.error) {
-    return res.status(400).json({ message: validationResult.error.message });
-  }
-
-  try {
-    const newContact = await addContact({ name, email, phone });
-    res.status(201).json(newContact);
-  } catch (error) {
-    console.error(error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-});
-
-app.delete('/api/contacts/:id', async (req, res) => {
-  const { id } = req.params;
-  try {
-    const result = await removeContact(id);
-    if (result) {
-      res.status(200).json({ message: 'Contact deleted' });
-    } else {
-      res.status(404).json({ message: 'Not found' });
-    }
-  } catch (error) {
-    console.error(error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-});
-
-app.put('/api/contacts/:id', async (req, res) => {
-  const { id } = req.params;
-  const { name, email, phone } = req.body;
-
-  const validationResult = validateContact({ name, email, phone });
-
-  if (validationResult.error) {
-    return res.status(400).json({ message: validationResult.error.message });
-  }
-
-  try {
-    const updatedContact = await updateContact(id, { name, email, phone });
-    if (updatedContact) {
-      res.status(200).json(updatedContact);
-    } else {
-      res.status(404).json({ message: 'Not found' });
-    }
-  } catch (error) {
-    console.error(error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-});
+app.use((err, req, res, next) => {
+  res.status(500).json({ message: err.message })
+})
 
 module.exports = app
