+++ models/contacts.js
@@ -1,14 +1,70 @@
-// const fs = require('fs/promises')
+const fs = require('fs/promises');
+const path = require('node:path');
 
-const listContacts = async () => {}
+const contactsPath = path.resolve('models', 'contacts.json');
 
-const getContactById = async (contactId) => {}
+const Joi = require('joi');
 
-const removeContact = async (contactId) => {}
+const contactSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+});
 
-const addContact = async (body) => {}
+function validateContact(contact) {
+  return contactSchema.validate(contact);
+}
 
-const updateContact = async (contactId, body) => {}
+const listContacts = async () => {
+  const data = await fs.readFile(contactsPath, 'utf-8');
+  return JSON.parse(data);
+}
+
+const getContactById = async (contactId) => {
+  const contacts = await listContacts();
+  return contacts.find((contact) => contact.id === contactId);
+}
+
+const removeContact = async (contactId) => {
+  const contacts = await listContacts();
+  const index = contacts.findIndex((contact) => contact.id === contactId);
+
+  if (index === -1) {
+    return null;
+  }
+
+  const [removedContact] = contacts.splice(index, 1);
+
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+
+  return removedContact;
+}
+
+const addContact = async (contact) => {
+  const contacts = await listContacts();
+  const newContact = { ...contact, id: String(Date.now()) };
+  contacts.push(newContact);
+
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+
+  return newContact;
+}
+
+const updateContact = async (contactId, body) => {
+  const contacts = await listContacts();
+  const index = contacts.findIndex((contact) => contact.id === contactId);
+
+  if (index === -1) {
+    return null;
+  }
+
+  const updatedContact = { ...contacts[index], ...body };
+  contacts[index] = updatedContact;
+
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+
+  return updatedContact;
+}
 
 module.exports = {
   listContacts,
@@ -16,4 +72,5 @@ module.exports = {
   removeContact,
   addContact,
   updateContact,
+  validateContact,
 }
