+++ models/contacts.json
@@ -1,11 +1,5 @@
 [
   {
-    "id": "AeHIrLTr6JkxGE6SN-0Rw",
-    "name": "Allen Raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
     "id": "qdggE76Jtbfd9eWJHrssH",
     "name": "Chaim Lewis",
     "email": "dui.in@egetlacus.ca",
@@ -58,5 +52,17 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
+  },
+  {
+    "name": "Chaim Bod",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685",
+    "id": "1701991903713"
+  },
+  {
+    "name": "Bod",
+    "email": "bob@egetlacus.ca",
+    "phone": "(294) 840-6685",
+    "id": "1701991916247"
   }
 ]
